backend/classcomunicator/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
deprecated==1.2.12
django-braces==1.14.0
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django==3.2.2
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jinja2==3.0.1
jwcrypto==0.9.1
markupsafe==2.0.1
model-bakery==1.3.1
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==8.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pyjwt[crypto]==2.1.0
pyparsing==2.4.7
pytest-django==4.3.0
pytest==6.2.4
pytz==2021.1
requests==2.25.1
ruamel.yaml.clib==0.2.4
ruamel.yaml==0.17.10
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
